Split PrintingConfig into different files
