+++ ObjectPrinting/IPropertyConfig.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Linq.Expressions;
+using ObjectPrinting.PrintingConfig;
 
 namespace ObjectPrinting
 {
